+++ frontend/src/components/__tests__/model-grid.test.tsx
@@ -1187,7 +1187,7 @@ describe("ModelBrowser", () => {
       await user.click(await screen.findByRole("button", { name: /Move/ }));
       const dialog = await screen.findByRole("dialog");
 
-      await user.click(within(dialog).getByRole("button", { name: /spares/ }));
+      await user.click(within(dialog).getByRole("button", { name: /Spares/ }));
       await user.click(within(dialog).getByRole("button", { name: /^Move/ }));
 
       await waitFor(() =>
@@ -2039,7 +2039,7 @@ describe("ModelBrowser", () => {
       await user.click(screen.getByLabelText("Select Benchy"));
       await user.click(await screen.findByRole("button", { name: /Move/ }));
       const dialog = await screen.findByRole("dialog");
-      await user.click(within(dialog).getByRole("button", { name: /spares/ }));
+      await user.click(within(dialog).getByRole("button", { name: /Spares/ }));
       await user.click(within(dialog).getByRole("button", { name: /^Move/ }));
     }